requirements.txt
openclaw>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.3.0
rank-bm25>=0.2.2
neo4j>=5.0.0
redis>=5.0.0
sqlite3-python
pyyaml>=6.0
pydantic>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.0.0
rich>=13.0.0
tiktoken>=0.5.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
